# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.MCPToolCallHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.MCPToolCallRecord'
        type: array
      total:
        type: integer
    type: object
  dto.MCPToolCallRecord:
    properties:
      duration:
        description: milliseconds
        type: integer
      error:
        type: string
      id:
        type: string
      parentMessageUUID:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      toolName:
        type: string
    type: object
...
  dto.RecordToolCallRequest:
    properties:
      duration:
        description: milliseconds
        minimum: 0
        type: integer
      error:
        type: string
      parent_message_uuid:
        type: string
      success:
        type: boolean
      toolName:
        type: string
    required:
    - duration
    - toolName
    type: object
...
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.SuccessResponseNoData:
    description: 成功的API响应标准格式,没有 data
    properties:
      code:
        example: 200
        type: integer
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: huahua1319873800@outlook.com
    name: API Support
    url: https://github.com/mingrenya/AI-Waf
  description: MRYa 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/mingrenya/AI-Waf
  title: MRYa-WAF API
  version: "1.0"
paths:
...
  /api/v1/mcp/tool-calls:
    get:
      consumes:
      - application/json
      description: 返回MCP工具的调用历史记录
      parameters:
      - default: 50
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MCPToolCallHistoryResponse'
              type: object
      summary: 获取MCP工具调用历史
      tags:
      - MCP
  /api/v1/mcp/tool-calls/record:
    post:
      consumes:
      - application/json
      description: MCP Server每次执行工具后调用此接口记录调用信息
      parameters:
      - description: 工具调用记录
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecordToolCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 记录MCP工具调用
      tags:
      - MCP
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"